+++ sdk/apimanagement/apimanagement-custom-widget-scaffolder/.eslintrc.json
@@ -1,0 +1,11 @@
+{
+  "plugins": [
+    "@azure/azure-sdk"
+  ],
+  "extends": [
+    "plugin:@azure/azure-sdk/azure-sdk-base"
+  ],
+  "rules": {
+    "@azure/azure-sdk/ts-package-json-files-required": "off"
+  }
+}
